Add `Button` component
* Install `Button` component (apps/client/src/components/ui/button.tsx);
* Remove old `Button` component (apps/client/src/components/ui/base/Button.tsx);
* Update `Header`, `LinkCreate`, `Login` & `Register` (apps/client/src/components/common/Header/Header.tsx, apps/client/src/components/template/LinkCreate/LinkCreate.tsx, apps/client/src/components/views/login/Login/Login.tsx & apps/client/src/components/views/register/Register/Register.tsx).
