+++ apps/client/src/components/views/register/Register/Register.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { Input } from "@client/components/ui/base/Input";
-import { Button } from "@client/components/ui/base/Button";
+import { Button } from "@client/components/ui/button";
 import { Typography } from "@client/components/ui/base/Typography";
 import { registerSchema, RegisterFormSchema } from "@client/types/auth";
 import { useForm, Controller } from "react-hook-form";
@@ -117,7 +117,7 @@ const Register = () => {
             </Typography>
           )}
         </div>
-        <Button type="submit">Register</Button>
+        <Button>Register</Button>
         {errors.registerError && (
           <Typography variant="body1" color="error">
             {errors.registerError.message}
